% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{read_xdrip_data}
\alias{read_xdrip_data}
\title{Read and Transform xDrip Data}
\usage{
read_xdrip_data(
  fil = here::here("Inst", "extdata", "Xdrip", "exportCSV20230518-182448.zip"),
  fil2 = here::here("Inst", "extdata", "Xdrip", "Sleep_times.xlsx")
)
}
\arguments{
\item{fil}{The path to the zipped .csv; defaults to what I used here}

\item{fil2}{The path to the sleep data; defaults to what I used here}
}
\value{
A data frame with columns 'DAY', 'TIME', 'UDT_CGMS', 'BG_LEVEL', 'REMARK', 'datetim', 'glucose', 
'remark_factor', and 'delta_time_remark' from the xDrip data, as well as transformed and grouped capillary glucose 
measurements, remarks, and factored remarks. 
Remarks and capillary glucose measurements are placed into the main data frame based on the closest timestamp.
}
\description{
This function reads in an xDrip data CSV file from a zip file, performs various transformations and wrangling
including calculating glucose levels in mmol/L units, joining with concomittant mean capillary glucose measurements, and joining the remarks into the data.
It also loads sleep data from an Excel file and factors remarks.
}
\examples{
data <- read_xdrip_data()
summary(data)

}
